Move search index collection to `utils/loading.py`
