+++ haystack/utils/app_loading.py
@@ -2,7 +2,6 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from django.apps import apps
-from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 
 __all__ = ['haystack_get_models', 'haystack_load_apps']
@@ -14,14 +13,6 @@ MODEL = 'model'
 def haystack_get_app_modules():
     """Return the Python module for each installed app"""
     return [i.module for i in apps.get_app_configs()]
-
-def haystack_get_search_indexes():
-    """Return all modules to collect search indexes in"""
-    app_modules = [(app_mod, '%s.search_indexes' % app_mod.__name__) for app_mod in haystack_get_app_modules()]
-    custom_search_indexes = getattr(settings, 'HAYSTACK_SEARCH_INDEX_MODULES', [])
-    app_modules.extend([(None, search_index) for search_index in custom_search_indexes])
-
-    return app_modules
 
 def haystack_load_apps():
     """Return a list of app labels for all installed applications which have models"""
