+++ test_haystack/test_app_loading.py
@@ -26,51 +26,6 @@ class AppLoadingTests(TestCase):
         for i in app_modules:
             self.assertIsInstance(i, ModuleType)
 
-    def test_get_search_indexes_with_unset_custom_search_indexes(self):
-        search_indexes = app_loading.haystack_get_search_indexes()
-        app_modules = app_loading.haystack_get_app_modules()
-
-        self.assertEqual(len(search_indexes), len(app_modules))
-        for i in app_modules:
-            self.assertIn((i, '%s.search_indexes' % i.__name__), search_indexes)
-
-    def test_get_search_indexes_with_empty_custom_search_indexes(self):
-        with self.settings(HAYSTACK_SEARCH_INDEX_MODULES=[]):
-            search_indexes = app_loading.haystack_get_search_indexes()
-        app_modules = app_loading.haystack_get_app_modules()
-
-        self.assertEqual(len(search_indexes), len(app_modules))
-        for i in app_modules:
-            self.assertIn((i, '%s.search_indexes' % i.__name__), search_indexes)
-
-    def test_get_search_indexes_with_custom_search_indexes(self):
-        SEARCH_INDEX_MODULES = [
-            'test_haystack.custom_search_indexes.custom_search_indexes',
-            'test_haystack.custom_search_indexes.custom_indexes',
-        ]
-
-        with self.settings(HAYSTACK_SEARCH_INDEX_MODULES=SEARCH_INDEX_MODULES):
-            search_indexes = app_loading.haystack_get_search_indexes()
-
-        for i in SEARCH_INDEX_MODULES:
-            self.assertIn((None, i), search_indexes)
-
-    def test_get_search_indexes_custom_search_indexes_extend_app_modules(self):
-        SEARCH_INDEX_MODULES = [
-            'test_haystack.custom_search_indexes.custom_search_indexes',
-            'test_haystack.custom_search_indexes.custom_indexes',
-        ]
-
-        with self.settings(HAYSTACK_SEARCH_INDEX_MODULES=SEARCH_INDEX_MODULES):
-            search_indexes = app_loading.haystack_get_search_indexes()
-            app_modules = app_loading.haystack_get_app_modules()
-
-        self.assertGreater(len(search_indexes), len(app_modules))
-        for i in SEARCH_INDEX_MODULES:
-            self.assertIn((None, i), search_indexes)
-        self.assertEqual(len(search_indexes),
-                         len(app_modules) + len(SEARCH_INDEX_MODULES))
-
     def test_get_models_all(self):
         models = app_loading.haystack_get_models('core')
         self.assertIsInstance(models, (list, GeneratorType))
