+++ test_haystack/test_loading.py
@@ -138,6 +138,53 @@ class ConnectionRouterTestCase(TestCase):
         self.assertEqual(cr.for_write(), ['multi1', 'multi2', 'default'])
 
 
+class CustomSearchIndexesTests(TestCase):
+    def test_get_search_indexes_with_unset_custom_search_indexes(self):
+        search_indexes = loading.haystack_get_search_indexes()
+        app_modules = loading.haystack_get_app_modules()
+
+        self.assertEqual(len(search_indexes), len(app_modules))
+        for i in app_modules:
+            self.assertIn((i, '%s.search_indexes' % i.__name__), search_indexes)
+
+    def test_get_search_indexes_with_empty_custom_search_indexes(self):
+        with self.settings(HAYSTACK_SEARCH_INDEX_MODULES=[]):
+            search_indexes = loading.haystack_get_search_indexes()
+        app_modules = loading.haystack_get_app_modules()
+
+        self.assertEqual(len(search_indexes), len(app_modules))
+        for i in app_modules:
+            self.assertIn((i, '%s.search_indexes' % i.__name__), search_indexes)
+
+    def test_get_search_indexes_with_custom_search_indexes(self):
+        SEARCH_INDEX_MODULES = [
+            'test_haystack.custom_search_indexes.custom_search_indexes',
+            'test_haystack.custom_search_indexes.custom_indexes',
+        ]
+
+        with self.settings(HAYSTACK_SEARCH_INDEX_MODULES=SEARCH_INDEX_MODULES):
+            search_indexes = loading.haystack_get_search_indexes()
+
+        for i in SEARCH_INDEX_MODULES:
+            self.assertIn((None, i), search_indexes)
+
+    def test_get_search_indexes_custom_search_indexes_extend_app_modules(self):
+        SEARCH_INDEX_MODULES = [
+            'test_haystack.custom_search_indexes.custom_search_indexes',
+            'test_haystack.custom_search_indexes.custom_indexes',
+        ]
+
+        with self.settings(HAYSTACK_SEARCH_INDEX_MODULES=SEARCH_INDEX_MODULES):
+            search_indexes = loading.haystack_get_search_indexes()
+            app_modules = loading.haystack_get_app_modules()
+
+        self.assertGreater(len(search_indexes), len(app_modules))
+        for i in SEARCH_INDEX_MODULES:
+            self.assertIn((None, i), search_indexes)
+        self.assertEqual(len(search_indexes),
+                         len(app_modules) + len(SEARCH_INDEX_MODULES))
+
+
 class MockNotAModel(object):
     pass
